Sorting implemented & working and files separated under package
