+++ Employee Management/bin/.gitignore
@@ -1,2 +1,21 @@
 /Employee.class
 /EmployeeManagement.class
+/EmployeeManagement$1.class
+/EmployeeManagement$2.class
+/EmployeeManagement$3.class
+/EmployeeSorting.class
+/EmployeeSorting$1.class
+/EmployeeSorting$2.class
+/EmployeeSorting$3.class
+/EmployeeSorting$4.class
+/EmployeeSorting$5.class
+/EmployeeSorting$6.class
+/DTO/
+/DAO/
+/LogicalAction$1.class
+/LogicalAction$2.class
+/LogicalAction$3.class
+/LogicalAction$4.class
+/LogicalAction$5.class
+/LogicalAction$6.class
+/LogicalAction.class
