+++ Employee Management/src/DTO/Employee.java
@@ -1,0 +1,46 @@
+package DTO;
+
+
+import java.time.LocalDate;
+
+public class Employee {
+
+    private int id;
+    private String name;
+    private String email;
+    private int age;
+    private LocalDate dob;
+
+    public int getId() {
+        return id;
+    }
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+    public void setName(String name) {
+        this.name = name;
+    }
+    public String getEmail() {
+        return email;
+    }
+    public void setEmail(String email) {
+        this.email = email;
+    }
+    public int getAge() {
+        return age;
+    }
+    public void setAge(int age) {
+        this.age = age;
+    }
+    public LocalDate getDob() {
+        return dob;
+    }
+    public void setDob(LocalDate dob) {
+        this.dob = dob;
+    }
+
+}
